pyproject.toml
[project]
name = "fractals"
version = "0.1.0"
description = "Render the classic fractals — escape-time, Newton, IFS, L-systems."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "dsculptor" }]
dependencies = [
    "numpy>=2.1",
    "pillow>=11.0",
    "typer>=0.15",
]

[project.scripts]
fractals = "fractals.cli:app"
build-all = "fractals.cli:build_all"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["fractals"]

[dependency-groups]
dev = [
    "mypy>=1.13",
    "pytest>=8.3",
    "ruff>=0.8",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
strict = true
python_version = "3.12"
files = ["fractals", "tests"]

# Improve readability of mypy output
show_error_codes = true
show_column_numbers = true
pretty = true

# Extra strictness on top of --strict
warn_unreachable = true
warn_unused_configs = true
enable_error_code = ["redundant-expr", "possibly-undefined", "truthy-bool"]

# Third-party packages that ship their own stubs (numpy ≥ 1.20, pillow-stubs,
# typer). If mypy is run outside the venv (e.g. by an IDE using system Python)
# it cannot locate those stubs — ignore rather than error on the import itself.
[[tool.mypy.overrides]]
module = [
    "numpy.*",
    "PIL.*",
]
ignore_missing_imports = true
